refactor: productUtils, cart 분리
