+++ src/basic/hooks/index.ts
@@ -1,3 +1,4 @@
 export * from './useDebounce';
 export * from './useNotification';
 export * from './useProduct';
+export * from './useCart';
